refactor: extract runtime reconciliation from AgentWorkspace
